fix: 면접 COMPLETE 상태 처리 로직 수정
