+++ src/components/common/_layout/InterviewLayout/InterviewLayout.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react'
 import { Outlet, useNavigate } from 'react-router-dom'
-import { Logo } from '@/components/common'
+import { Button, Logo, Modal } from '@/components/common'
 import { Timer } from '@/components/interview'
 import { useInterviewStore } from '@/stores/interviewStore'
 import { getInterviewRouteByPhase } from '@/utils'
@@ -10,6 +10,8 @@ export const InterviewLayout: React.FC = () => {
   const { isInterviewer, currentPhase, questionOption, selectedQuestion } =
     useInterviewStore()
   const navigate = useNavigate()
+
+  const isLastRoundDone = currentPhase.toUpperCase() === 'COMPLETE'
 
   useEffect(() => {
     if (!isInterviewer) {
@@ -38,6 +40,15 @@ export const InterviewLayout: React.FC = () => {
       <div className={styles.pageContent}>
         <Outlet />
       </div>
+
+      <Modal
+        isOpen={isLastRoundDone}
+        closeOnBgClick={false}
+        style="congrats"
+        message={['오늘의 면접이 모두 종료되었습니다', '수고하셨습니다.']}
+      >
+        <Button text="홈으로 이동" onClick={() => navigate('/')} />
+      </Modal>
     </div>
   )
 }
