+++ src/pages/interview/InterviewAwaitingPage/InterviewAwaitingPage.tsx
@@ -47,15 +47,8 @@ export const InterviewAwaitingPage: React.FC = () => {
 
   const { startTimer, resetTimer } = useTimerStore()
   const { getOddInterviewee, getEvenInterviewee } = useMatchStore()
-  const {
-    interviewId,
-    isInterviewer,
-    currentRound,
-    currentPhase,
-    setInterviewData,
-  } = useInterviewStore()
-
-  const isLastRoundDone = currentPhase.toUpperCase() === 'COMPLETE'
+  const { interviewId, isInterviewer, currentRound, setInterviewData } =
+    useInterviewStore()
 
   const { data: matchResult } = useMatchResult(false)
   const { data: currentStatus, refetch } = useInterviewStatus(interviewId)
@@ -251,15 +244,6 @@ export const InterviewAwaitingPage: React.FC = () => {
           '면접 상대가 정해지지 않았습니다.',
           '1:1 매칭을 먼저 진행해주세요.',
         ]}
-      >
-        <Button text="홈으로 이동" onClick={() => navigate('/')} />
-      </Modal>
-
-      <Modal
-        isOpen={isLastRoundDone}
-        closeOnBgClick={false}
-        style="congrats"
-        message={['오늘의 면접이 모두 종료되었습니다', '수고하셨습니다.']}
       >
         <Button text="홈으로 이동" onClick={() => navigate('/')} />
       </Modal>
